+++ mlir/utils/performance/conv-configs
