+++ mlir/utils/performance/createPerformanceReports.py
@@ -8,53 +8,47 @@ import pandas as pd
 import sys
 
 #Create html reports from .csv files
-def printAllPerformance(chip):
-    def readReport(file, kind):
-        try:
-            return pd.read_csv(chip + '_' + file)
-        except FileNotFoundError:
-            print(f'${kind} report not found.')
-            return None
+def printAllPerformance(chip, lib='rocBLAS'):
+    perfReportFound = False
 
-    COLUMNS_TO_AVERAGE = ['MLIR TFlops', 'MIOpen TFlops (no MLIR Kernels)', 'MLIR/MIOpen']
-
-    df = readReport(reportUtils.PERF_REPORT_FILE, 'Perf')
-    if df.empty:
+    try:
+        df = pd.read_csv(chip + '_' + reportUtils.PERF_REPORT_FILE[lib])
+        perfReportFound = True
+        if 'Tuned MLIR TFlops' in df:
+            COLUMNS_TO_AVERAGE = ['MLIR TFlops', 'Tuned MLIR TFlops',
+                f'{lib} TFlops (no MLIR Kernels)', 'Tuned/Untuned',
+                f'MLIR/{lib}', f'Tuned/{lib}']
+        else:
+            COLUMNS_TO_AVERAGE = ['MLIR TFlops',
+                f'{lib} TFlops (no MLIR Kernels)',
+                f'MLIR/{lib}']
+    except FileNotFoundError:
+        print('Perf report not found.')
         return
-    tuned_df = readReport(reportUtils.MIOPEN_TUNED_REPORT_FILE,
-                          'MIOpen with tuned MLIR')
-    untuned_df = readReport(reportUtils.MIOPEN_UNTUNED_REPORT_FILE,
-                            'MIOpen with untuned MLIR')
 
-    # Add tuned and untuned performance to the existing performance table
-    if not tuned_df.empty and not untuned_df.empty:
-        df['MIOpen TFlops (Tuned MLIR Kernels)'] = tuned_df['TFlops']
-        df['MIOpen TFlops (Untuned MLIR Kernels)'] = untuned_df['TFlops']
-        df['Tuned/Untuned'] = df['MIOpen TFlops (Tuned MLIR Kernels)']/df['MIOpen TFlops (Untuned MLIR Kernels)']
-        df['Tuned/MIOpen'] = df['MIOpen TFlops (Tuned MLIR Kernels)']/df['MIOpen TFlops (no MLIR Kernels)']
-        df.to_csv(chip + '_' + reportUtils.PERF_REPORT_FILE, index=False)
-        COLUMNS_TO_AVERAGE = ['MLIR TFlops', 'MIOpen TFlops (no MLIR Kernels)', 'MLIR/MIOpen',
-                              'MIOpen TFlops (Tuned MLIR Kernels)', 'MIOpen TFlops (Untuned MLIR Kernels)',
-                              'Tuned/Untuned', 'Tuned/MIOpen']
-
-    plotMean = df[COLUMNS_TO_AVERAGE].agg(reportUtils.geoMean)
+    # Only plot the actual averages, not the ratios
+    # (This conveniently keeps the old behavior for the no tuning DB case)
+    plotMean = df[COLUMNS_TO_AVERAGE[:3]].agg(reportUtils.geoMean)
     plotMean.name = "Geo. mean"
     plotMean = pd.DataFrame(plotMean).T
 
-    plotMean[['MLIR TFlops', 'MIOpen TFlops (no MLIR Kernels)']]\
-        .to_csv(chip + '_' + reportUtils.PERF_PLOT_REPORT_FILE, index=False)
+    plotMean[['MLIR TFlops', f'{lib} TFlops (no MLIR Kernels)']]\
+        .to_csv(chip + '_' + reportUtils.PERF_PLOT_REPORT_FILE[lib], index=False)
 
-    means = df.groupby(["Direction", "DataType", "InputLayout"])[COLUMNS_TO_AVERAGE]\
+    means = df.groupby(["DataType"])[COLUMNS_TO_AVERAGE]\
         .agg(reportUtils.geoMean)
-    means.loc[("All", "All", "All"),:] = df[COLUMNS_TO_AVERAGE].agg(reportUtils.geoMean)
-    means.to_csv(chip + '_' + reportUtils.PERF_STATS_REPORT_FILE)
+    means.loc["All"] = df[COLUMNS_TO_AVERAGE].agg(reportUtils.geoMean)
+    means.to_csv(chip + '_' + reportUtils.PERF_STATS_REPORT_FILE[lib])
 
-    with open(chip + "_" + "MLIR_vs_MIOpen.html", 'w') as htmlOutput:
-        reportUtils.htmlReport(df, means, "MLIR vs. MIOpen performance",
-          ["MLIR/MIOpen", "Tuned/Untuned", "Tuned/MIOpen"],
-          reportUtils.colorForSpeedups,
-          htmlOutput)
+    toHighlight = [f"MLIR/{lib}"]
+    if "Tuned/Untuned" in df:
+        toHighlight += [f"Tuned/{lib}", f"Tuned/Untuned"]
+
+    with open(chip + "_" + f"MLIR_vs_{lib}.html", 'w') as htmlOutput:
+        reportUtils.htmlReport(df, means, f"MLIR vs. {lib} performance",
+        toHighlight, reportUtils.colorForSpeedups, htmlOutput)
 
 # Main function.
 if __name__ == '__main__':
-    printAllPerformance(sys.argv[1])
+    lib = sys.argv[2] if len(sys.argv) > 2 else 'rocBLAS'
+    printAllPerformance(sys.argv[1], lib)
